Split Sparkle preflight test helpers
