+++ Packages/CmuxUpdater/Tests/CmuxUpdaterTests/CapturingUpdateLog.swift
@@ -1,0 +1,15 @@
+import Foundation
+@testable import CmuxUpdater
+
+final class CapturingUpdateLog: UpdateLogging, @unchecked Sendable {
+    private let lock = NSLock()
+    private(set) var messages: [String] = []
+
+    func append(_ message: String) {
+        lock.withLock {
+            messages.append(message)
+        }
+    }
+
+    func logPath() -> String { "/tmp/cmux-update-test.log" }
+}
